src/test/java/org/nikiforova/solutions/easy/array/ContainsDuplicateTest.java
package org.nikiforova.solutions.easy.array;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

class ContainsDuplicateTest {
    private static Stream<Arguments> provideParameters() {
        return Stream.of(
                Arguments.of(new int[]{1, 2, 3, 1}, 3, true),
                Arguments.of(new int[]{1, 0, 1, 1}, 1, true),
                Arguments.of(new int[]{1, 2, 3, 1, 2, 3}, 2, false)
        );
    }

    @ParameterizedTest
    @MethodSource("provideParameters")
    @DisplayName("Test containsNearbyDuplicate with different arguments")
    void containsNearbyDuplicate(int[] nums, int k, boolean expected) {
        assertEquals(expected, ContainsDuplicate.containsNearbyDuplicate(nums, k));
    }
}